add main page, move budget to budget route
