+++ app/globals.css
@@ -31,3 +31,9 @@ body {
     text-wrap: balance;
   }
 }
+
+@layer components {
+  .btn-primary {
+    @apply font-bold border-2 border-black rounded-md p-2;
+  }
+}
